requirements.txt
streamlit==1.36.0
streamlit-image-comparison==0.0.4
torch==2.3.1
torchvision==0.18.1
ultralytics==8.2.52
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.3.0
PyMuPDF==1.24.5
transformers==4.41.2
sentencepiece==0.2.0
accelerate==0.31.0